server/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.3
reportlab==4.0.7
Pillow==10.2.0

